Improved Stream writer/reader
commit 4cfeb7ad5fd7ee2520134190af36396f6d620bba
Author: lriki <lriki.net@gmail.com>
Date:   Wed Jun 29 20:14:43 2022 +0900

    wip

commit 8dcb63e538b12ab2335706280b909244f3fd9619
Author: lriki <lriki.net@gmail.com>
Date:   Wed Jun 29 18:22:10 2022 +0900

    wip

commit 094d283c0f712eef92fc1027a1db4efb771b3a1a
Author: lriki <lriki.net@gmail.com>
Date:   Wed Jun 29 16:54:37 2022 +0900

    wip

commit e00683f8941b39ba0bfc809e8fd90a938505da2c
Author: lriki <lriki.net@gmail.com>
Date:   Wed Jun 29 13:54:54 2022 +0900

    wip

commit 6a07b3f9750fe687dde73cbdf5df69d1010d6540
Author: lriki <lriki.net@gmail.com>
Date:   Wed Jun 29 13:34:18 2022 +0900

    wip

commit ffc1394fc59584f2db23f78c5d2a34a4e44ee3f4
Author: lriki <lriki.net@gmail.com>
Date:   Wed Jun 29 12:45:41 2022 +0900

    wip

commit 2eb4a5bc0e3958981929d73d6b9c42f3e9115b3e
Author: lriki <lriki.net@gmail.com>
Date:   Wed Jun 29 12:31:48 2022 +0900

    wip

+++ lumino/LuminoCore/src/IO/FileSystem.cpp
@@ -22,7 +22,7 @@ namespace ln {
 class PlatformFileFinder
     : public RefObject {
 public:
-    PlatformFileFinder(const PathChar* dirPath, int dirPathLen, FileAttribute attr, const PathChar* pattern, int patternLen);
+    PlatformFileFinder(const PathChar* dirPath, size_t dirPathLen, FileAttribute attr, const PathChar* pattern, int patternLen);
     ~PlatformFileFinder();
     bool isWorking() const;
     bool isFirst() const;
@@ -44,7 +44,7 @@ private:
     bool m_first;
 };
 
-PlatformFileFinder::PlatformFileFinder(const PathChar* dirPath, int dirPathLen, FileAttribute attr, const PathChar* pattern, int patternLen)
+PlatformFileFinder::PlatformFileFinder(const PathChar* dirPath, size_t dirPathLen, FileAttribute attr, const PathChar* pattern, int patternLen)
     : m_impl()
     , m_dirPath(dirPath)
     , m_current()
